+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="index.css">
     <title>Awesome books</title>
   </head>
   <body>
@@ -88,7 +88,7 @@
         2023 &copy Copyright
       </p>
     </footer>
-    <script type="module" src="script.js" defer></script>
+    <script type="module" src="./index.js" defer></script>
   </div>
   </body>
 </html>
