+++ modules/library.js
@@ -1,6 +1,6 @@
 import Book from './book.js';
 
-class Library {
+export default class Library {
   constructor() {
     this.library = document.querySelector('.library');
     this.titleInput = document.querySelector('#title');
@@ -17,7 +17,7 @@ class Library {
     });
   }
 
-  addBook() {
+  addBook = () => {
     const title = this.titleInput.value;
     const author = this.authorInput.value;
     const newBook = new Book(title, author);
@@ -27,11 +27,11 @@ class Library {
     this.displayLibrary();
   }
 
-  updateLocalStorage() {
+  updateLocalStorage = () => {
     localStorage.setItem('books', JSON.stringify(this.books));
   }
 
-  displayLibrary() {
+  displayLibrary = () => {
     this.library.innerHTML = '';
     this.books.forEach((book, index) => {
       const bookElement = document.createElement('li');
@@ -59,48 +59,9 @@ class Library {
     });
   }
 
-  removeBook(book) {
+  removeBook = (book) => {
     this.books = this.books.filter((b) => b !== book);
     this.updateLocalStorage();
     this.displayLibrary();
   }
 }
-
-const library = new Library();
-library.displayLibrary();
-
-const links = document.querySelectorAll('.nav-link');
-const sections = document.querySelectorAll('.section');
-
-links.forEach((link, index) => {
-  link.addEventListener('click', (e) => {
-    e.preventDefault();
-    sections.forEach((section) => {
-      if (!section.classList.contains('hidden')) {
-        section.classList.add('hidden');
-      }
-    });
-    links.forEach((link) => {
-      if (!link.classList.contains('color-black')) {
-        link.classList.add('color-black');
-      }
-    });
-    sections[index].classList.remove('hidden');
-    links[index].classList.remove('color-black');
-  });
-});
-
-function currentDate() {
-  const dataAndTime = new Date();
-  const year = dataAndTime.getFullYear();
-  const day = dataAndTime.getDate();
-  const hours = dataAndTime.getHours().toString().padStart(2, '0');
-  const minutes = dataAndTime.getMinutes().toString().padStart(2, '0');
-  const monthName = new Intl.DateTimeFormat('en-US', { month: 'long' }).format(dataAndTime);
-  return `${monthName} ${day}th ${year}, ${hours}:${minutes}`;
-}
-
-document.addEventListener('DOMContentLoaded', () => {
-  const datCon = document.querySelector('.date');
-  datCon.innerHTML = currentDate();
-});
