+++ package.json
@@ -269,6 +269,7 @@
     "longest-streak": "^2.0.4",
     "lowercase-keys": "^2.0.0",
     "lru-cache": "^7.18.3",
+    "luxon": "^3.3.0",
     "make-dir": "^3.1.0",
     "make-fetch-happen": "^11.1.0",
     "map-age-cleaner": "^0.1.3",
